+++ .eslintrc.js
@@ -1,0 +1,14 @@
+module.exports = {
+    "env": {
+        "node": true,
+        "commonjs": true,
+        "es6": true
+    },
+    "extends": "eslint:recommended",
+    "parserOptions": {
+        "ecmaVersion": "latest"
+    },
+    "rules": {
+        "no-empty": ["error", { "allowEmptyCatch": true }]
+    }
+}