+++ devices/simple-switch.js
@@ -19,10 +19,12 @@ class SimpleSwitch extends TuyaDevice {
             }
         }
 
-        // Send home assistant discovery data and give it a second before sending state updates
-        this.initDiscovery()
-        await utils.sleep(1)
-
+        if (this.publish_homeassistant_discovery) {
+            // Send home assistant discovery data and give it a second before sending state updates
+            this.initDiscovery()
+            await utils.sleep(1)
+        }
+        
         // Get initial states and start publishing topics
         this.getStates()
     }
@@ -32,9 +34,9 @@ class SimpleSwitch extends TuyaDevice {
 
         const discoveryData = {
             name: (this.config.name) ? this.config.name : this.config.id,
-            state_topic: this.baseTopic+'state',
-            command_topic: this.baseTopic+'command',
-            availability_topic: this.baseTopic+'status',
+            state_topic: this.options.baseTopic+'state',
+            command_topic: this.options.baseTopic+'command',
+            availability_topic: this.options.baseTopic+'status',
             payload_available: 'online',
             payload_not_available: 'offline',
             unique_id: this.config.id,
@@ -43,7 +45,7 @@ class SimpleSwitch extends TuyaDevice {
 
         debugDiscovery('Home Assistant config topic: '+configTopic)
         debugDiscovery(discoveryData)
-        this.publishMqtt(configTopic, JSON.stringify(discoveryData))
+        this.publishMqtt({topic: configTopic, message: JSON.stringify(discoveryData)})
     }
 }