+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Forked from upstream at Feb 2022. The changelist below reflects changes vs. orig
 - TBD
 
 ## Changed
-- TBD
+- Dependencies updated to latest-available (as of 2/6/2022)
 
 ## [3.0.0]
 The 3.0.0 release is a major refactor of the project with significant changes from previous version.  Only major additions and changes are listed below.